updated car field types and constructor to better match db

+++ MPG Tracker/src/com/unrulyrecursion/mpgtracker/data/CarList.java
@@ -36,7 +36,7 @@ public class CarList {
 
     private static void addItem(Car item) {
         ITEMS.add(item.getCarName());
-        ITEM_MAP.put(Integer.toString(item.getId()), item.getCarName());
+        ITEM_MAP.put(Long.toString(item.getId()), item.getCarName());
     }
